# Enhanced Kernel Dispatch Flexibility in SlangPy