… slangpy/slang/callshape.slang
… slangpy/tests/device/test_buffer.py
… slangpy/tests/slangpy_tests/generated_tests/read_slice_generic_error.slang
… slangpy/tests/slangpy_tests/nested_types_generics.slang
… slangpy/tests/slangpy_tests/test_call_group_integrations.py
… slangpy/tests/slangpy_tests/test_call_groups.py
… src/slangpy_ext/utils/slangpy.cpp
… src/slangpy_ext/utils/slangpy.h