Enhanced Kernel Dispatch Flexibility in SlangPy (#72)

This commit introduces enhanced kernel dispatch flexibility via
a new "call_group_shape()" function that enables more fine-grained
control over GPU thread organization and dispatch patterns. A simple
usage example might look something like:

`call_data = module.test2D.call_group_shape(Shape((4,8)))(spy.grid((32,64)), spy.thread_id(), _result='numpy')`

Where a shader named "test2D" is being given a call group shape of
4x8 (note that the ordering is [y,x] here), a call shape of 32x64
as a result of the spy.grid((32,64)) being passed in, as well as given
the thread_id. The mapping of call-id's to thread-id's from such a call
would look like:

[call-id[0] (y), call-id[1] (x), thread-id]

[   0    0    0]
[   1    0    1]
[   2    0    2]
...
[  61    0  229]
[  62    0  230]
[  63    0  231]
[   0    1    8]
[   1    1    9]
[   2    1   10]
...
[  61    1  237]
[  62    1  238]
[  63    1  239]

Running without the call_group_shape function:

`call_data = module.test2D.(spy.grid((32,64)), spy.thread_id(), _result='numpy')`

or alternatively:

`call_data = module.test2D.call_group_shape(Shape(()))(spy.grid((32,64)), spy.thread_id(), _result='numpy')`

results in SlangPy using it's standard linear dispatch scheme using call groups
containing 32 threads and the mapping from call-id's to thread-id's from such
calls would look like:

[   0    0    0]
[   1    0    1]
[   2    0    2]
...
[  61    0   61]
[  62    0   62]
[  63    0   63]
[   0    1   64]
[   1    1   65]
[   2    1   66]
...
[  61    1  125]
[  62    1  126]
[  63    1  127]

Conceptually, the default when not specifying a call group shape can be thought
of as running dispatches where each thread has its own call group, but in
actuallity, threads are still run in groups of 32. Thus a call like:

`call_data = module.test2D_thread.call_group_shape(Shape((1,1)))(spy.grid((32,64)), spy.thread_id(), _result='numpy')`

Will ultimately function differently than the two examples above and would likely
have worse performance. Similarly:

`call_data = module.test2D_thread.call_group_shape(Shape((1,32)))(spy.grid((32,64)), spy.thread_id(), _result='numpy')`

Can also function differently from the default case where call group shape is
not specified, as we can now run into alignment issues when the call shape is
not aligned to 32 threads in the "x" dimension (ex a 1x33 call group shape on a
32x64 call shape).

The new functionality also works with arbitrary dimensionalities:

`...test5D_thread.call_group_shape(Shape((2,2,2,2,2)))(spy.grid((2,2,2,32,64)), spy.thread_id(), _result='numpy')`

Using the new call_group_shape functionality, developers can now specify how
threads are arranged spatially, allowing for better uses of shared memory and
better control over memory access patterns.

In order to support the new functionality, the change extends the existing
dispatch system to support call_group_shape parameters with proper validation,
padding, and alignment handling, while also handling edge cases for misaligned
shapes and dimensions smaller than group sizes, while also providing error reporting
for invalid configurations.

The change includes unit tests that validate call group functionality across
1D through 5D dispatches to ensure mathematical consistency of the group ID
calculations and thread ID assignments. The tests cover both typical use
cases and edge cases and have been added under:

-slangpy/tests/slangpy_tests/test_call_groups.py

The change also includes integration tests to ensure call group functionality
works with SlangPy's existing major features such as tensor operations with
autodiff support, NDBuffer processing, texture sampling and filtering, and
transformation mapping operations. These integration tests have been added
under:

-slangpy/tests/slangpy_tests/test_call_group_integrations.py