@sricker-nvidia 2025-06-10
This PR is almost ready to be moved out of the "draft" state, I still need to double check some of the testing, remove some of the personal test files, updated the commit message etc.

I was able to test this against https://github.com/elcuppietan/slangpy-samples/tree/wave-ballot/experiments/balloted-splatting using the attached patch:
[call_group_shape_edits.patch](https://github.com/user-attachments/files/20664876/call_group_shape_edits.patch)

and things seem to work as expected (or at least run the same as the original sample). That said I don't notice a performance difference. I'm not sure if that's expected or not.