@sricker-nvidia 2025-06-11
Noting that I see some warnings in the CI:
````
slangpy/tests/slangpy_tests/test_call_group_integrations.py ....::warning file=/opt/ci/actions-runner/_work/slangpy/slangpy/slangpy/tests/slangpy_tests/helpers.py,line=168::Conversion of an array with ndim > 0 to a scalar is deprecated, and will error in future. Ensure you extract a single element from your array before performing this operation. (Deprecated NumPy 1.25.)
Warning: Conversion of an array with ndim > 0 to a scalar is deprecated, and will error in future. Ensure you extract a single element from your array before performing this operation. (Deprecated NumPy 1.25.)
````