@sricker-nvidia 2025-06-12
Resolved the review comments that I could. I added another perf test under https://github.com/sricker-nvidia/slangpy/tree/call-id-improve-perf-test/slangpy/tests/call_groups that just tests the default linear case for the sake of being able to compare the PR against ToT. This is what some of the results looked like:

default only perf test
	-Note that this seems faster than the default case in call_group_performance_test.py
	 likely because get_call_id<2>() is slow and it is not used here (spy.call_id used instead)
	-PR
		Baseline Performance: 3.589 ± 0.369 ms
		Baseline Performance: 3.535 ± 0.305 ms
		Baseline Performance: 3.463 ± 0.046 ms
		Baseline Performance: 3.494 ± 0.188 ms
		Baseline Performance: 3.548 ± 0.319 ms

	-ToT
		Baseline Performance: 3.524 ± 0.211 ms
		Baseline Performance: 3.522 ± 0.173 ms
		Baseline Performance: 3.464 ± 0.083 ms
		Baseline Performance: 3.519 ± 0.223 ms
		Baseline Performance: 3.542 ± 0.289 ms

ToT seems to wins out slightly because in the PR we still populate call_group_id and call_group_thread_id,even in the default case leading to a small perf hit. However, my most recent runs are showing around 3.680 ± 0.440 ms, so it seems like I may have made an edit somewhere that reduced things back a bit. It could also just be noise/variance.

I also revisited the balloted-splatting example and had some interesting observations:
ballot splat perf
	-with PR
		-no use of call group shapes (default linear dispatch)
			-27 it/sec
		-using call group shapes
			-3.17it/sec
	-ToT
		-no use of call group shapes (default linear dispatch)
			-27 it/sec

So it looks like I was mistaken when previously I thought I saw no difference in perf in that sample, the call groups seem to be making things slower, even though the sample still has correct output. I don't fully understand what the sample is doing, so there could be something wrong with the edits in my patch adding in call groups.