@kaizhangNV 2025-06-18
I think the perf issue is just because we use the NDBuffer<float, 2> instead of 2D texture as the input.
1×32 groups will win, because it has the most cache friendly access pattern.
NDBuffer is actually still 1-dimensional buffer though its name is `ND`. But texture is different because the memory layout is designed to cache a 2/3-D block in the hardware.