@sricker-nvidia 2025-06-25
Did some digging today. The perf issue on the balloted-splatting example seems to be the result of the newly added call-id calculations in the generated kernel code.

I tested by rewriting parts of the PR such that everything functions as before, but get_call_id will still return the newly calculated, non linear call id based on the given call group shape. That allowed me to tinker with using the pre computed call-ids and the call group shape call ids in the same function call. Using the value from get_call_id() gives the correct visual result, but slows things down. If the value from get_call_id() is used in a way that can be optimized out by the compiler, things speed up.

I haven't thought too deeply about possible solutions yet. One might be to try and pre calculate dispatch id's in NativeCallData::exec and send them in as a uniform, but that seems like it could have it's own perf impacts. Another idea is to somehow store the id's that threads calculate and reuse those if the same function call is made with the same call group shape. The most realistic idea I think of is to implement what was suggest on the issue, "We could also optimize where appropriate (or when hinted to by the user), and map call groups directly to thread groups where possible, which would simply reduce the need for arithmetic operations". I would expect that to get perf back on par, but would only be a good solution for up to 3D calls, but maybe that's ok for now?

Perf should in theory get worse with higher dimensionalities due to additional arithmetic operations required for each additional dimension. Not sure if there's a good way around that unless ids are precalculated and reused in multiple passes.