@sricker-nvidia 2025-06-27
Uploaded another update patch integrating changes from https://github.com/shader-slang/slangpy/compare/kai/dev/call-shape-link-time (thanks @kaizhangNV !). The perf issue on the balloted-splatting example has been resolved and was due to dynamic indexing in some for loops. Using [unroll] to force static indexing fixed the issue.