@sricker-nvidia 2025-06-27
Also uploading a new patch for the balloted-splatting example using call group shapes
[call_group_shape.patch](https://github.com/user-attachments/files/20957473/call_group_shape.patch)

Performance for ToT balloted-splatting example with ToT slangpy: runs in 05:55
Performance for balloted-splatting example using call group shapes and current version of the PR: runs in 06:01

For additional comparison the updates_1 patch ran in 5:57